+++ src/pytask_parallel/plugin.py
@@ -8,7 +8,6 @@ from pytask import hookimpl
 
 from pytask_parallel import build
 from pytask_parallel import config
-from pytask_parallel import execute
 from pytask_parallel import logging
 
 if TYPE_CHECKING:
@@ -20,5 +19,4 @@ def pytask_add_hooks(pm: PluginManager) -> None:
     """Register plugins."""
     pm.register(build)
     pm.register(config)
-    pm.register(execute)
     pm.register(logging)
