+++ src/pytask_parallel/utils.py
@@ -1,0 +1,61 @@
+"""Contains utility functions."""
+
+from __future__ import annotations
+
+import inspect
+from typing import TYPE_CHECKING
+from typing import Any
+
+from pytask.tree_util import PyTree
+from pytask.tree_util import tree_leaves
+from pytask.tree_util import tree_map
+from pytask.tree_util import tree_structure
+
+from pytask_parallel.backends import ParallelBackend
+
+if TYPE_CHECKING:
+    from pytask import PTask
+
+
+def create_kwargs_for_task(task: PTask) -> dict[str, PyTree[Any]]:
+    """Create kwargs for task function."""
+    parameters = inspect.signature(task.function).parameters
+
+    kwargs = {}
+    for name, value in task.depends_on.items():
+        kwargs[name] = tree_map(lambda x: x.load(), value)
+
+    for name, value in task.produces.items():
+        if name in parameters:
+            kwargs[name] = tree_map(lambda x: x.load(), value)
+
+    return kwargs
+
+
+def handle_task_function_return(task: PTask, out: Any) -> None:
+    """Handle the return value of a task function."""
+    if "return" not in task.produces:
+        return
+
+    structure_out = tree_structure(out)
+    structure_return = tree_structure(task.produces["return"])
+    # strict must be false when none is leaf.
+    if not structure_return.is_prefix(structure_out, strict=False):
+        msg = (
+            "The structure of the return annotation is not a subtree of "
+            "the structure of the function return.\n\nFunction return: "
+            f"{structure_out}\n\nReturn annotation: {structure_return}"
+        )
+        raise ValueError(msg)
+
+    nodes = tree_leaves(task.produces["return"])
+    values = structure_return.flatten_up_to(out)
+    for node, value in zip(nodes, values):
+        node.save(value)
+
+
+def is_parallelized(n_workers: int, parallel_backend: ParallelBackend) -> bool:
+    """Check if the execution is parallelized."""
+    return parallel_backend == ParallelBackend.CUSTOM or (
+        n_workers > 1 and parallel_backend != ParallelBackend.CUSTOM
+    )
