+++ tests/test_execute.py
@@ -12,13 +12,11 @@ from pytask_parallel.execute import _Sleeper
 
 from tests.conftest import restore_sys_path_and_module_after_test_execution
 
-
-class Session:
-    pass
+_IMPLEMENTED_BACKENDS = [p for p in ParallelBackend if p != ParallelBackend.CUSTOM]
 
 
 @pytest.mark.end_to_end()
-@pytest.mark.parametrize("parallel_backend", ParallelBackend)
+@pytest.mark.parametrize("parallel_backend", _IMPLEMENTED_BACKENDS)
 def test_parallel_execution(tmp_path, parallel_backend):
     source = """
     from pytask import Product
@@ -40,7 +38,7 @@ def test_parallel_execution(tmp_path, parallel_backend):
 
 
 @pytest.mark.end_to_end()
-@pytest.mark.parametrize("parallel_backend", ParallelBackend)
+@pytest.mark.parametrize("parallel_backend", _IMPLEMENTED_BACKENDS)
 def test_parallel_execution_w_cli(runner, tmp_path, parallel_backend):
     source = """
     from pytask import Product
@@ -70,7 +68,7 @@ def test_parallel_execution_w_cli(runner, tmp_path, parallel_backend):
 
 
 @pytest.mark.end_to_end()
-@pytest.mark.parametrize("parallel_backend", ParallelBackend)
+@pytest.mark.parametrize("parallel_backend", _IMPLEMENTED_BACKENDS)
 def test_stop_execution_when_max_failures_is_reached(tmp_path, parallel_backend):
     source = """
     import time
@@ -98,7 +96,7 @@ def test_stop_execution_when_max_failures_is_reached(tmp_path, parallel_backend)
 
 
 @pytest.mark.end_to_end()
-@pytest.mark.parametrize("parallel_backend", ParallelBackend)
+@pytest.mark.parametrize("parallel_backend", _IMPLEMENTED_BACKENDS)
 def test_task_priorities(tmp_path, parallel_backend):
     source = """
     import pytask
@@ -139,7 +137,7 @@ def test_task_priorities(tmp_path, parallel_backend):
 
 
 @pytest.mark.end_to_end()
-@pytest.mark.parametrize("parallel_backend", ParallelBackend)
+@pytest.mark.parametrize("parallel_backend", _IMPLEMENTED_BACKENDS)
 @pytest.mark.parametrize("show_locals", [True, False])
 def test_rendering_of_tracebacks_with_rich(
     runner, tmp_path, parallel_backend, show_locals
@@ -221,7 +219,7 @@ def test_sleeper():
 
 
 @pytest.mark.end_to_end()
-@pytest.mark.parametrize("parallel_backend", ParallelBackend)
+@pytest.mark.parametrize("parallel_backend", _IMPLEMENTED_BACKENDS)
 def test_task_that_return(runner, tmp_path, parallel_backend):
     source = """
     from pathlib import Path
@@ -241,7 +239,7 @@ def test_task_that_return(runner, tmp_path, parallel_backend):
 
 
 @pytest.mark.end_to_end()
-@pytest.mark.parametrize("parallel_backend", ParallelBackend)
+@pytest.mark.parametrize("parallel_backend", _IMPLEMENTED_BACKENDS)
 def test_task_without_path_that_return(runner, tmp_path, parallel_backend):
     source = """
     from pathlib import Path
@@ -263,7 +261,7 @@ def test_task_without_path_that_return(runner, tmp_path, parallel_backend):
 
 @pytest.mark.end_to_end()
 @pytest.mark.parametrize("flag", ["--pdb", "--trace", "--dry-run"])
-@pytest.mark.parametrize("parallel_backend", ParallelBackend)
+@pytest.mark.parametrize("parallel_backend", _IMPLEMENTED_BACKENDS)
 def test_parallel_execution_is_deactivated(runner, tmp_path, flag, parallel_backend):
     tmp_path.joinpath("task_example.py").write_text("def task_example(): pass")
     result = runner.invoke(
@@ -277,7 +275,12 @@ def test_parallel_execution_is_deactivated(runner, tmp_path, flag, parallel_back
 @pytest.mark.end_to_end()
 @pytest.mark.parametrize("code", ["breakpoint()", "import pdb; pdb.set_trace()"])
 @pytest.mark.parametrize(
-    "parallel_backend", [i for i in ParallelBackend if i != ParallelBackend.THREADS]
+    "parallel_backend",
+    [
+        i
+        for i in ParallelBackend
+        if i not in (ParallelBackend.THREADS, ParallelBackend.CUSTOM)
+    ],
 )
 def test_raise_error_on_breakpoint(runner, tmp_path, code, parallel_backend):
     tmp_path.joinpath("task_example.py").write_text(f"def task_example(): {code}")
@@ -289,7 +292,7 @@ def test_raise_error_on_breakpoint(runner, tmp_path, code, parallel_backend):
 
 
 @pytest.mark.end_to_end()
-@pytest.mark.parametrize("parallel_backend", ParallelBackend)
+@pytest.mark.parametrize("parallel_backend", _IMPLEMENTED_BACKENDS)
 def test_task_partialed(runner, tmp_path, parallel_backend):
     source = """
     from pathlib import Path
